+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Updated unit tests according to above fix
 * Load model only when frame dimensions change
 * Fixed crash from openGL due to potential race condition.
+* Fixed videos paused when change from other tabs to video tab and repen app from backgroud
+* Fixed screen sharing is unavaiable to present videos
 
 ## [0.17.1] - 2022-06-03