+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Updated unit tests according to above fix
 * Load model only when frame dimensions change
 * Fixed crash from openGL due to potential race condition.
+* [Demo] Fixed videos paused when change from other tabs to video tab and reopen app from background
+* [Demo] Fixed screen sharing is unavailable to present videos
 
 ## [0.17.1] - 2022-06-03